+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.instagram_app",
+    package = "com.jays_space.photosphere",
 )
 
 android_resource(
     name = "res",
-    package = "com.instagram_app",
+    package = "com.jays_space.photosphere",
     res = "src/main/res",
 )
