+++ android/app/src/main/java/com/jays_space/photosphere/newarchitecture/modules/MainApplicationTurboModuleManagerDelegate.java
@@ -1,4 +1,4 @@
-package com.instagram_app.newarchitecture.modules;
+package com.jays_space.photosphere.newarchitecture.modules;
 
 import com.facebook.jni.HybridData;
 import com.facebook.react.ReactPackage;
