+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">instagram_app</string>
+    <string name="app_name">Jays PhotoSphere</string>
 </resources>
