+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "instagram_app",
-  "displayName": "instagram_app"
+  "name": "JaysPhotoSphere",
+  "displayName": "Jays PhotoSphere"
 }
