+++ ios/JaysPhotoSphere.xcodeproj/xcshareddata/xcschemes/JaysPhotoSphere.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "instagram_app.app"
-               BlueprintName = "instagram_app"
-               ReferencedContainer = "container:instagram_app.xcodeproj">
+               BuildableName = "JaysPhotoSphere.app"
+               BlueprintName = "JaysPhotoSphere"
+               ReferencedContainer = "container:JaysPhotoSphere.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "instagram_appTests.xctest"
-               BlueprintName = "instagram_appTests"
-               ReferencedContainer = "container:instagram_app.xcodeproj">
+               BuildableName = "JaysPhotoSphereTests.xctest"
+               BlueprintName = "JaysPhotoSphereTests"
+               ReferencedContainer = "container:JaysPhotoSphere.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "instagram_app.app"
-            BlueprintName = "instagram_app"
-            ReferencedContainer = "container:instagram_app.xcodeproj">
+            BuildableName = "JaysPhotoSphere.app"
+            BlueprintName = "JaysPhotoSphere"
+            ReferencedContainer = "container:JaysPhotoSphere.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "instagram_app.app"
-            BlueprintName = "instagram_app"
-            ReferencedContainer = "container:instagram_app.xcodeproj">
+            BuildableName = "JaysPhotoSphere.app"
+            BlueprintName = "JaysPhotoSphere"
+            ReferencedContainer = "container:JaysPhotoSphere.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
