+++ ios/JaysPhotoSphere/AppDelegate.h
