+++ ios/JaysPhotoSphere/AppDelegate.mm
