+++ ios/JaysPhotoSphere/Images.xcassets/AppIcon.appiconset/Contents.json
