+++ ios/JaysPhotoSphere/Images.xcassets/Contents.json
