+++ ios/JaysPhotoSphereTests/Info.plist
