+++ ios/JaysPhotoSphereTests/JaysPhotoSphereTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface instagram_appTests : XCTestCase
+@interface JaysPhotoSphereTests : XCTestCase
 
 @end
 
-@implementation instagram_appTests
+@implementation JaysPhotoSphereTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
