+++ ios/Podfile
@@ -5,7 +5,7 @@ require_relative '../node_modules/@react-native-community/cli-platform-ios/nativ
 platform :ios, '12.0'
 install! 'cocoapods', :deterministic_uuids => false
 
-target 'instagram_app' do
+target 'JaysPhotoSphere' do
   use_expo_modules!
   post_integrate do |installer|
     begin
@@ -28,7 +28,7 @@ target 'instagram_app' do
     :app_path => "#{Pod::Config.instance.installation_root}/.."
   )
 
-  target 'instagram_appTests' do
+  target 'JaysPhotoSphereTests' do
     inherit! :complete
     # Pods for testing
   end
