+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "instagram_app",
+  "name": "JaysPhotoSphere",
   "version": "0.0.1",
   "private": true,
   "scripts": {
